Factor audio engine interop out of `Clip`
Currently the `Clip` entity is the main complexity bottleneck of
the system. Much of this complexity, especially code specific to
web audio APIs, needs to be hoisted up a level to allow multiple
Clips to be orchestrated into a gapless playback queue.

This commit preps for that work, pulling the playback state and
audio engine specifics out of the `Clip` and into their own
modules / constructs.